feat(bulma): transform project using bulma
