+++ .browserslistrc
@@ -1,0 +1,2 @@
+> 10%
+last 10 versions
