+++ assets/css/objects/index.css
@@ -1,0 +1,3 @@
+.wrapper {
+  padding: 30px;
+}
