+++ assets/css/settings/colors.css
@@ -1,0 +1,3 @@
+:root {
+  --white: #fff;
+}
