+++ assets/css/settings/index.css
@@ -1,0 +1,1 @@
+@import "./colors.css";
