--- assets/css/style.css
@@ -1,141 +1,0 @@
-body {
-    background-color: #ecf0f1;
-    padding: 30px;
-}
-
-.main {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.box {
-    background-color: #fff;
-    border-radius: 4px;
-    max-width: 600px;
-    padding: 30px;
-    box-shadow: 0px 0px 20px 0px rgba(0,0,0,0.22);
-    margin-bottom: 50px;
-}
-
-.title {
-    color: #2c3e50;
-    font-size: 1.5em;
-    text-align: center;
-    margin-bottom: 40px;
-    font-weight: bold;
-}
-
-.input-group {
-    margin-bottom: 25px;
-}
-
-.input-label {
-    font-size: 0.9em;
-    display: block;
-    color: #2c3e50;
-    margin-bottom: 7px;
-}
-
-.input-element {
-    font-size: 0.8em;
-    padding: 10px;
-    width: 100%;
-    border: 1px solid #bdc3c7;
-    border-radius: 5px;
-}
-
-.button {
-    width: 100%;
-    border: 0;
-    border-radius: 5px;
-    padding: 20px 10px;
-    font-size: 0.9em;
-    margin: 5px 0;
-}
-
-.button.primary {
-    background-color: #3498db;
-    color: #fff;
-}
-
-#generated-image {
-    width: 500px;
-    height: 550px;
-    background-image: url("../images/background.png");
-    padding: 30px;
-}
-
-.border {
-    height: 100%;
-    border: 4px solid #fff;
-    padding: 20px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-evenly;
-}
-
-.message {
-    font-family: 'Dancing Script', cursive;
-    color: #fff;
-    text-align: center;
-    font-size: 2em;
-    margin-bottom: 20px;
-    width: 100%;
-}
-
-.preacher-image {
-    width: 130px;
-    height: 130px;
-    border-radius: 100px;
-    margin-bottom: 10px;
-}
-
-.preacher {
-    color: #fff;
-    font-family: 'Lato', sans-serif;
-    text-align: center;
-    font-weight: 700;
-    text-transform: uppercase;
-    font-size: 1em;
-    margin-bottom: 10px;
-    width: 100%;
-}
-
-.message-date {
-    color: #fff;
-    font-family: 'Lato', sans-serif;
-    text-align: center;
-    font-weight: 400;
-    font-size: 1em;
-    width: 100%;
-}
-
-.church-address {
-    color: #fff;
-    font-family: 'Lato', sans-serif;
-    text-align: center;
-    font-weight: 300;
-    font-size: 0.8em;
-    width: 100%;
-}
-
-.church-logo {
-    max-width: 70px;
-    margin-bottom: 10px;
-    margin-top: 20px;
-}
-
-.generated-image-footer {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.generated-image-preacher {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
