+++ assets/js/pages/index.js
@@ -1,0 +1,6 @@
+/* eslint-disable */
+import preacher from './pregacao'
+
+export {
+  preacher,
+}
