+++ postcss.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  plugins: {
+    autoprefixer: {},
+    'postcss-import': {},
+    'postcss-nested': {},
+    cssnano: {},
+  },
+}
